Reorg bulk
